+++ mod/code/datalist_row_id_cell_template.php
@@ -1,5 +1,5 @@
             <a class="edit" href="javascript:void(0)">
                 <div class="w3-col {{COLUMN_WIDTH}} w3-padding-small">
-                    <span class="w3-tag w3-theme">{{id}}</span>
+                    <span class="w3-tag <?php echo $color['tag']; ?>">{{id}}</span>
                 </div>
             </a>